[TASK] Add ControllerContext while rendering AbstractMessage
The ControllerContext contains a Request formatted with the ExtensionName of the subclass of the AbstractMessage being sent - i.e. if you subclassed to Tx_MyExt_Message_SuperDuperMessage or MyCompany\MyExt\Message\SuperDuperMessage, the "MyExt" part is used as ExtensionName in the Request.

This allows ViewHelpers like f:translate to correctly fetch LLL from the extension containing the specific Message implementation rather than from Notify itself.